+++ opositores/opositores.php
@@ -70,7 +70,6 @@
 	<title>Web oficial de Aproahp - Opositores</title>
 </head>
 <body>
-
 <div id="container"><!-- inicio container -->
 <div id="mensaje" style="display:none" class="ui-widget-content ui-corner-all">
 	<h3 class="ui-widget-header ui-corner-all">Status</h3><p><?php echo $Mensaje ?></p>
@@ -78,22 +77,21 @@
 <?php require("../header.php"); ?>
 	<div id="izda-a">
 		<div class="resalte">
-		<p><strong>En esta página puedes encontrar información de interés.
+		<h1>Opositores</h1>
+		<p>En esta página puedes encontrar información de interés.
 		si estás preparándote para obtener una plaza de
 		Agente de la Hacienda Pública
-		tanto por el turno libre como por el de promoción interna.</strong></p>
+		tanto por el turno libre como por el de promoción interna.</p>
 		</div>
-				
 		<div id='opositores'>
-		<h1>Opositores</h1>
 		<form  id='opositoresform' method='post' action='<?php echo $paginaactual ?>'>
 			<table>
 				<tr><td class='label1form'><label for='inp_titulo'>Título:</label></td><td><input type='text' name='inp_titulo' id='inp_titulo' maxlength='50'/></td></tr>
 				<tr><td class='label1form'><label for='inp_autor'>Autor:</label></td><td><input type='text' name='inp_autor' id='inp_autor' maxlength='50'/></td></tr>
 				<tr><td class='label1form'><label for='inp_email'>Email:</label></td><td><input type='text' name='inp_email' id='inp_email' maxlength='50'/>&nbsp;(Opcional, no será mostrado a otros usuarios)</td></tr>
 				<tr><td colspan='2'><label for='inp_comentario'>Comentario:</label></td></tr>
-				<tr><td colspan='2'><TEXTAREA style="width:600px;height:180px;" name='inp_comentario' id='inp_comentario' maxlength='5000'></TEXTAREA></td></tr>
-				<tr><td colspan='2'><div id='captchawidget'><?php echo recaptcha_get_html($captchaPublicK); ?></div><input class='botonsubmit' type='submit' value='Enviar' name='agregarcomentario'/></td></tr>
+				<tr><td colspan='2'><textarea name='inp_comentario' id='inp_comentario' maxlength='5000'></textarea></td></tr>
+				<tr><td colspan='2'><div id='captchawidget'><?php echo recaptcha_get_html($captchaPublicK); ?></div><input class='botonsubmit' type='submit' value='Enviar Formulario' name='agregarcomentario'/></td></tr>
 			</table>
 		</form>
 <?php
@@ -121,7 +119,6 @@
 ?>
 		</div><!-- fin div opositores-->
 	</div>
-
 	<div id="dcha-a">
 			<div class="box-blue">
 				<h2 class="section"><b></b></h2>